// src/features/list/errors-enum.ts
export enum ErrorList {
  NotFound = "Not found",
  Forbidden = "Forbidden",
  AlreadyExists = "Ya existe una lista con ese ID",
  MaxListsExceeded = "No puedes crear más de 3 listas",
  Unauthorized = "No autorizado",
  InvalidInput = "Entrada inválida",
  DatabaseError = "Error de base de datos",
  ServerError = "Error del servidor",
}